update and fix

+++ application/app/Http/Controllers/Backend/EventController.php
@@ -101,7 +101,7 @@ class EventController extends Controller
 	            }
 		    }
         }
-
+        Session::flash('success', 'Data Has Been Added');
     	return redirect()->route('admin.event');
     }
 
@@ -208,32 +208,58 @@ class EventController extends Controller
         {
             EventImage::destroy($request->remove_image);
         }
-        
+        Session::flash('success', 'Data Has Been Updated');
     	return redirect()->route('admin.event');
     }
 
     public function delete($id)
     {
     	Event::destroy($id);
-
+        Session::flash('success', 'Data Has Been Deleted');
     	return redirect()->route('admin.event');
     }
 
     public function action(Request $request)
     {
-    	if($request->action == 'delete')
-    	{
-    		Event::destroy($request->id);
-    	}
-    	else if($request->action == 'enable')
-    	{
-    		$index = Event::whereIn('id', $request->id)->update(['flag_publish' => 1]);
-    	}
-    	else if($request->action == 'disable')
-    	{
-    		$index = Event::whereIn('id', $request->id)->update(['flag_publish' => 0]);
-    	}
+        if(isset($request->id))
+        {
+        	if($request->action == 'delete')
+        	{
+        		Event::destroy($request->id);
+                Session::flash('success', 'Data Selected Has Been Deleted');
+        	}
+        	else if($request->action == 'enable')
+        	{
+        		$index = Event::whereIn('id', $request->id)->update(['flag_publish' => 1]);
+                Session::flash('success', 'Data Selected Has Been Enabled');
+        	}
+        	else if($request->action == 'disable')
+        	{
+        		$index = Event::whereIn('id', $request->id)->update(['flag_publish' => 0]);
+                Session::flash('success', 'Data Selected Has Been Disabled');
+        	}
+        }
     	
     	return redirect()->route('admin.event');
+    }
+
+    public function publish($id, $action)
+    {
+        $index = Event::find($id);
+
+        $index->flag_publish = $action;
+
+        $index->save();
+
+        if($action == 1)
+        {
+            Session::flash('success', 'Data Has Been Enabled');
+        }
+        else
+        {
+            Session::flash('success', 'Data Has Been Disabled');
+        }
+
+        return redirect()->route('admin.event');
     }
 }
