+++ application/app/Http/Controllers/Backend/GalleryController.php
@@ -93,6 +93,7 @@ class GalleryController extends Controller
 
         $index->save();
 
+        Session::flash('success', 'Data Has Been Added');
     	return redirect()->route('admin.gallery');
     }
 
@@ -183,31 +184,58 @@ class GalleryController extends Controller
 
         $index->save();
 
+        Session::flash('success', 'Data Has Been Updated');
     	return redirect()->route('admin.gallery');
     }
 
     public function delete($id)
     {
     	Gallery::destroy($id);
-
+        Session::flash('success', 'Data Has Been Deleted');
     	return redirect()->route('admin.gallery');
     }
 
     public function action(Request $request)
     {
-    	if($request->action == 'delete')
-    	{
-    		Gallery::destroy($request->id);
-    	}
-    	else if($request->action == 'enable')
-    	{
-    		$index = Gallery::whereIn('id', $request->id)->update(['flag_publish' => 1]);
-    	}
-    	else if($request->action == 'disable')
-    	{
-    		$index = Gallery::whereIn('id', $request->id)->update(['flag_publish' => 0]);
-    	}
+        if(isset($request->id))
+        {
+        	if($request->action == 'delete')
+        	{
+        		Gallery::destroy($request->id);
+                Session::flash('success', 'Data Selected Has Been Deleted');
+        	}
+        	else if($request->action == 'enable')
+        	{
+        		$index = Gallery::whereIn('id', $request->id)->update(['flag_publish' => 1]);
+                Session::flash('success', 'Data Selected Has Been Enabled');
+        	}
+        	else if($request->action == 'disable')
+        	{
+        		$index = Gallery::whereIn('id', $request->id)->update(['flag_publish' => 0]);
+                Session::flash('success', 'Data Selected Has Been Disabled');
+        	}
+        }
     	
     	return redirect()->route('admin.gallery');
+    }
+
+    public function publish($id, $action)
+    {
+        $index = Gallery::find($id);
+
+        $index->flag_publish = $action;
+
+        $index->save();
+
+        if($action == 1)
+        {
+            Session::flash('success', 'Data Has Been Enabled');
+        }
+        else
+        {
+            Session::flash('success', 'Data Has Been Disabled');
+        }
+
+        return redirect()->route('admin.gallery');
     }
 }
